+++ src/Ark.jl
@@ -4,6 +4,9 @@ using FieldViews
 using Preferences
 using StaticArrays
 using FunctionWrappers: FunctionWrapper
+using Random
+
+const THREAD_SAFE_LOCK = @load_preference("THREAD_SAFE_LOCK", default = "true")
 
 isdefined(@__MODULE__, :Memory) || const Memory = Vector # Compat for Julia < 1.11
 
@@ -34,6 +37,7 @@ include("observer.jl")
 include("filter.jl")
 include("query.jl")
 include("batch_ops.jl")
+include("unchecked.jl")
 !_is_testing() && include("precompile.jl")
 
 #include("docs.jl") # doctest setup
@@ -51,7 +55,7 @@ export Entity
 export is_zero
 
 export Query, Filter
-export close!, count_entities, unregister!
+export close!, count_entities, shuffle_entities!, unregister!
 
 export Entities
 
@@ -60,7 +64,10 @@ export OnCreateEntity, OnRemoveEntity, OnAddComponents, OnRemoveComponents
 export OnAddRelations, OnRemoveRelations
 
 export Observer, observe!, emit_event!
+
 export unpack, @unpack
+
+export @unchecked
 
 export Relationship