+++ test/TestTypes.jl
@@ -41,6 +41,10 @@ struct NoIsBits
     v::Vector{Int}
 end
 
+struct NoIsBits2
+    v::Vector{Vector{Int}}
+end
+
 mutable struct MutableNoIsBits
     v::Vector{MutableComponent}
 end
@@ -51,4 +55,13 @@ struct ChildOf3 <: Relationship end
 
 mutable struct Tick
     time::Int
+end
+
+struct Position_Mod
+    x::Float64
+    y::Float64
+end
+
+function Base.getproperty(value::Position_Mod, name::Symbol)
+    return nothing
 end