+++ test/test_lock.jl
@@ -1,34 +1,32 @@
 
-@testset "_Lock functionality" begin
+@testset "_Lock basic functionality" begin
     lock = _Lock()
 
     # Initially, nothing should be locked
     @test !_is_locked(lock)
-    @test lock.lock_bits == 0
+    @test lock._counter == 0
 
-    # Lock a bit
-    b1 = _lock(lock)
+    # Lock one time
+    _lock(lock)
     @test _is_locked(lock)
-    @test ((lock.lock_bits >> (b1 - 1)) & 0x01) == 1
+    @test lock._counter == 1
 
-    # Lock another bit
-    b2 = _lock(lock)
-    @test b2 != b1
-    @test ((lock.lock_bits >> (b2 - 1)) & 0x01) == 1
-
-    # Unlock first bit
-    _unlock(lock, b1)
-    @test ((lock.lock_bits >> (b1 - 1)) & 0x01) == 0
-    @test _is_locked(lock)  # still locked because b2 is active
-
-    # Unlock second bit
-    _unlock(lock, b2)
+    # Unlock one time
+    _unlock(lock)
+    @test lock._counter == 0
     @test !_is_locked(lock)
-    @test lock.lock_bits == 0
 
-    # Unlocking an already unlocked bit should throw
-    @test_throws(
-        "InvalidStateException: unbalanced unlock. Did you close a query that was already iterated?",
-        _unlock(lock, b1)
-    )
+    for _ in 1:2
+        Threads.@sync begin
+            Threads.@spawn for _ in 1:10^3
+                _lock(lock)
+            end
+            Threads.@spawn for _ in 1:10^3
+                _unlock(lock)
+            end
+        end
+    end
+
+    @test lock._counter == 0
+    @test !_is_locked(lock)
 end