+++ CHANGELOG.md
@@ -9,12 +9,22 @@
 ### Features
 
 - Adds support for user-defined storage types (#437)
-- Adds a new type GPUVector which can be used to offload work to the GPU (#470, #476)
+- Adds new types `GPUVector` and `GPUStructArray` which can be used to offload work to the GPU (#470, #476, #483)
+- Makes the locking mechanism thread-safe for parallel queries (#536, #537)
+- Adds support for shuffling tables based on filters (#527)
 - Adds support for UniqueVector storage (#481)
+
+### Performance
+
+- Add an unchecked macro to trade-off safety for speed (#521, #522, #531)
 
 ### Bugfixes
 
 - Fix copy on special mutable types (#515, fixes #514)
+
+### Other
+
+- Closing an already closed query is a no-op instead of raising an error (#493)
 
 ## [[v0.3.2]](https://github.com/ark-ecs/Ark.jl/compare/v0.3.1...v0.3.2)